// src/types.ts
export interface ChatMessage {
  id: string;
  role: 'user' | 'model';
  content: string;
  timestamp: Date;
  isLoading?: boolean;
  citations?: Citation[];
}

export interface Citation {
  id: string;
  title: string; // ex: "PPA 2026-2029 - Pág 45"
  uri: string; // Link para o arquivo
  snippet: string; // Trecho do texto
}